+++ packages/core/src/calculators/index.ts
@@ -1,6 +1,7 @@
 export * from "./common.js";
 export * from "./apgar.js";
 export * from "./silvermanAndersen.js";
+export * from "./woodDownesFerres.js";
 export * from "./flacc.js";
 export * from "./qtc.js";
 export * from "./schwartz.js";